Backend/requirements.txt
# FastAPI Core & Infrastructure
fastapi==0.129.0
uvicorn==0.41.0
pydantic==2.12.5
pydantic_core==2.41.5
email-validator==2.3.0
anyio==4.12.1
starlette==0.52.1
websockets==16.0
python-multipart==0.0.6

# MongoDB
pymongo==4.16.0
dnspython==2.8.0

# Authentication & Security
python-jose==3.5.0
bcrypt==4.0.1
passlib==1.7.4
argon2-cffi==25.1.0
cryptography==46.0.5

# Environment Variables
python-dotenv==1.0.0

# Utilities
requests==2.32.5
pytz==2025.2
python-dateutil==2.9.0.post0
typing_extensions==4.15.0

# AI Proctoring & Computer Vision
opencv-python==4.13.0.92
opencv-contrib-python==4.13.0.92
numpy==2.2.6
ultralytics==8.4.14
torch==2.10.0
torchvision==0.25.0
pillow==12.1.1
scipy==1.15.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
sympy==1.14.0

# Google Gemini generative AI
google-generativeai>=0.1.0

# Supplemental & Transitive
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
ecdsa==0.19.1
exceptiongroup==1.3.1
filelock==3.24.2
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
google-auth==2.48.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
packaging==26.0
polars==1.38.1
polars-runtime-32==1.38.1
protobuf==3.20.3
psutil==7.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
tzdata==2025.3
ultralytics-thop==2.0.18
urllib3==2.6.3
watchfiles==1.1.1

# HTTP client and testing suites
httpx>=0.24.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
